requirements.txt
discord.py==2.3.2
PyNaCl>=1.5.0
groq>=0.11.0
qrcode[pil]>=7.4.2
yt-dlp>=2024.7.25

